% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anonymize_randomly.R
\name{anonymize_randomly}
\alias{anonymize_randomly}
\title{Generate a random string for each unique value in a variable from a data frame}
\usage{
anonymize_randomly(mydf, identifier, id_length = 10, drop_old = TRUE)
}
\arguments{
\item{mydf}{A data frame or data table}

\item{identifier}{A variable in the data that contains the characters or numbers you would like to replace with an anonymous ID. Should be wrapped in quotes.}

\item{id_length}{The length of the newly generated random id. Defaults to 10.}

\item{drop_old}{Do you want to drop the identifier from the data that gets returned? Defaults to TRUE.}
}
\value{
A data frame identical to the original but with a new variable containing the random ID, which will be named <identifier>.anon.ID, and missing the identifier if drop_old is set to TRUE.
}
\description{
Generate a random string for each unique value in a variable from a data frame
}
\examples{
anonymize_randomly(mydata, "Social.Security", 9, FALSE)
}

