% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbation}
\alias{perturbation}
\alias{randomNoise}
\title{Add Random Noise to Data}
\usage{
perturbation(x, cols = 1:ncol(x), magnitude = "MEDIUM", minNoise = NULL,
  maxNoise = NULL, minNoiseQuantile = NULL, maxNoiseQuantile = NULL)

randomNoise(x, cols = 1:ncol(x), magnitude = "MEDIUM", minNoise = NULL,
  maxNoise = NULL, minNoiseQuantile = NULL, maxNoiseQuantile = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{cols}{the columns that are to be affected. Must be numeric.
If no value is given, all columns are affected.}

\item{magnitude}{a shortcut for the potential size of noise - see details}

\item{minNoise}{the smallest possible value for noise that may be added.
If neither minNoise nor minNoiseQuantile is provided, minNoise defaults to 0.}

\item{maxNoise}{the largest possible value for noise that may be added}

\item{minNoiseQuantile}{the smallest possible value for noise that may be added.
Rather than an actual value, this is a decimal, representing a quantile of the
values in that column.
If both minNoise and minNoiseQuantile are provided, maxNoise takes priority.}

\item{maxNoiseQuantile}{the largest possible value for noise that may be added.
Rather than an actual value, this is a decimal, representing a quantile of the
values in that column.
If both maxNoise and maxNoiseQuantile are provided, maxNoise takes priority.}
}
\value{
data.frame with with noise added all elements in specified columns
}
\description{
Add Random Noise to Data
}
\details{
Adds a random value between minNoise and maxNoise to each element in each of the
specified columns.
If minNoiseQuantile or maxNoiseQuantile is provided instead of minNoise or
maxNoise, then minNoise and maxNoise are calculated for every column based on
the quantiles provided.
magnitude is a shortcut for minNoise and maxNoiseQuantile parameters:
\itemize{
     \item LOW:      minNoise = 0   maxNoiseQuantile = .25
     \item MEDIUM:   minNoise = 0   maxNoiseQuantile = .5
     \item HIGH:     minNoise = 0   maxNoiseQuantile = .75
}
This makes magnitude ideal when the user desires to have a noise value between
0 and some quantile for all specified columns.
}
\section{Functions}{
\itemize{
\item \code{randomNoise}: identical to perturbation
}}
\examples{
perturbation(data)
perturbation(data, 1:5)
perturbation(data, magnitude="HIGH")
perturbation(data, "Gender", minNoise = 10, maxNoise = 50)
perturbation(data, minNoiseQuantile = .1, maxNoise = .9)
}

