% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomizeCharacters.R
\name{randomizeCharacters}
\alias{randomizeCharacters}
\title{Replace Data with Random Character Strings}
\usage{
randomizeCharacters(x, cols = 1:ncol(x), randomizeNames = FALSE,
  consistent = TRUE, numChars = 10)
}
\arguments{
\item{x}{a data.frame}

\item{cols}{the columns that are to be replaced - may be numeric (column indices)
or character (column names). If no value is given, all columns are replaced.}

\item{randomizeNames}{whether or not column names should be randomized as well}

\item{consistent}{whether or not reocurring values continue to hold the same value as
each other after they are randomized}

\item{numChars}{the number of characters the randomized data should contain}
}
\value{
data.frame with with specified columns replaced by random strings
}
\description{
Replace Data with Random Character Strings
}
\details{
Replaces all values within the specified columns with random 10-length
character strings.
By default, if a value reoccurs within the column, that value will be assigned
the same random character string each time. Column names are also randomized.
Both of these options may be changed.
}
\examples{
randomCharacters(data)
randomCharacters(data, 2, consistent=FALSE)
randomCharacters(data, c("Gender","Education"), randomizeNames=TRUE)
}

