% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/salt.R
\name{salt}
\alias{salt}
\title{Salt a vector.}
\usage{
salt(.x, .chars = letters, .n_chars = 5L)
}
\arguments{
\item{.x}{a vector.}

\item{.chars}{set of characters to salt with.}

\item{.n_chars}{an integer; number of characters to salt with.}
}
\value{
A salted version of the vector.
}
\description{
Salt a vector.
}
\examples{
set.seed(1)
# Use various number of characters
salt(letters, .n_chars = 0L)
salt(letters, .n_chars = 1L)
salt(letters, .n_chars = 5L)
salt(letters)

# Use other sets of characters to salt with
salt(letters, .chars = letters[1:2])
}

