% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize a vector.}
\usage{
anonymize(.x, .algo = "sha256")
}
\arguments{
\item{.x}{a vector.}

\item{.algo}{the name of the algorithm.}
}
\value{
An anonymized version of the data object.
}
\description{
Anonymize a vector.
}

