% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anonymizer.R
\docType{package}
\name{anonymizer}
\alias{anonymizer}
\alias{anonymizer-package}
\title{anonymizer: Anonymize Data Containing Personally Identifiable Information}
\description{
anonymizer: Anonymize Data Containing Personally Identifiable Information
}

