% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deanonymize.R
\name{deanonymize}
\alias{deanonymize}
\title{Deanonymize a vector.}
\usage{
deanonymize(.x, .algo = "sha256", .seed = 0, .chars = letters,
  .n_chars = 5L, ...)
}
\arguments{
\item{.x}{a vector.}

\item{.algo}{the name of the algorithm.}

\item{.seed}{an integer to seed the random number generator.}

\item{.chars}{set of characters to unsalt with.}

\item{.n_chars}{an integer; number of characters to unsalt with.}

\item{...}{additional arguments to be based to \code{unhash}.}
}
\value{
An deanonymized version of the vector.
}
\description{
Deanonymize a vector.
}

