% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unsalt.R
\name{unsalt}
\alias{unsalt}
\title{Unsalt a vector.}
\usage{
unsalt(.x, .seed = 0)
}
\arguments{
\item{.x}{a vector.}

\item{.seed}{an integer to seed the random number generator.}
}
\value{
An unsalted version of the vector.
}
\description{
Unsalt a vector.
}

